+++ src/scenes/movement-test-scene.ts
@@ -1,6 +1,6 @@
 import 'phaser';
 
-import { SystemsManager } from '../lib/systems-manager';
+import { SystemsManager } from '../lib/systems/systems-manager';
 import { SignSystem } from '../systems/sign-system';
 import { HasSpriteSystem } from '../systems/has-sprite-system';
 import { HasPhysicalSpriteSystem } from '../systems/has-physical-sprite-system';
