+++ src/systems/sign-system.ts
@@ -1,5 +1,5 @@
-import { BaseSystem } from '../lib/base-system';
-import { SystemsManager } from '../lib/systems-manager';
+import { BaseSystem } from '../lib/systems/base-system';
+import { SystemsManager } from '../lib/systems/systems-manager';
 
 type Sign = Systems.HasIndicator.Entity & Systems.HasInteractionCircle.Entity & Systems.HasSprite.Entity;
